Josh is awesome :)
"Hey,

I may have done this for no reason other than for myself... but I went
ahead and linked all of the lighter color to that swatch and cleaned
the defs. Hopefully it will save you guys a little time or energy. At
the very least, it looks more consistent for my use. :)"